NumberSprite Update
The gfx info on NumberSprites was ugly, taking up way more space than it
needed to.  This update shrinks the size of NumberSprites and makes the
according adjustment in the tabs so that the NumberSprites appear in the
same positions.  This update works only with NDSRAF as of the
NumberSprite Update in NDSRAF.